# backend/src/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ReqAI – AI Requirement Analyzer API
  version: 1.0.0
  description: |
    Enterprise REST API for ReqAI — AI-powered requirement analysis platform.

    ## Authentication
    All protected endpoints require a Bearer JWT in the `Authorization` header.
    Obtain tokens via `POST /auth/login`. Refresh via `POST /auth/refresh` (httpOnly cookie).

    ## Response Envelope
    All responses follow a consistent envelope:
    ```json
    { "success": true, "data": {}, "meta": { "page": 1, "limit": 20, "total": 142 } }
    ```
    Errors:
    ```json
    { "success": false, "error": { "code": "VALIDATION_ERROR", "message": "...", "details": [] } }
    ```

  contact:
    name: ReqAI Engineering
    email: engineering@reqai.dev
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.reqai.dev/v1
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token (15-minute expiry). Obtain via POST /auth/login.

  # ─── Reusable Parameters ──────────────────────────────────────────────────────
  parameters:
...
    AnalysisId:
      name: analysisId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  # ─── Reusable Responses ──────────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authentication required
...
    Conflict:
      description: Resource already exists or state conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Request body failed schema validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'

  # ─── Schemas ─────────────────────────────────────────────────────────────────
  schemas:
...
    UserRole:
      type: string
      enum: [ADMIN, BUSINESS_ANALYST, DEVELOPER, QA_ENGINEER, ARCHITECT, PROJECT_MANAGER]
...
    AnalysisStatus:
      type: string
      enum: [QUEUED, PROCESSING, COMPLETED, FAILED, CANCELLED]
    ComplexityLevel:
      type: string
      enum: [LOW, MEDIUM, HIGH, VERY_HIGH]
...
    AIProvider:
      type: string
      enum: [OPENAI, AZURE_OPENAI, ANTHROPIC, WATSONX, CUSTOM]
...
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:    { type: string, example: NOT_FOUND }
            message: { type: string, example: Requirement not found }
    ValidationErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:    { type: string, example: VALIDATION_ERROR }
            message: { type: string, example: Request validation failed }
            details:
              type: array
              items:
                type: object
                properties:
                  field:   { type: string, example: email }
                  message: { type: string, example: Invalid email address }
    PaginationMeta:
      type: object
      properties:
        page:       { type: integer, example: 1 }
        limit:      { type: integer, example: 20 }
        total:      { type: integer, example: 142 }
        totalPages: { type: integer, example: 8 }
        hasNext:    { type: boolean, example: true }
        hasPrev:    { type: boolean, example: false }

    # ── Auth Schemas ──────────────────────────────────────────────────────────
    RegisterRequest:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email:      { type: string, format: email, example: ananya@company.com }
        password:
          type: string
          minLength: 8
          description: Must contain uppercase, number, and special character
          example: "MyP@ssw0rd!"
        firstName:  { type: string, example: Ananya }
        lastName:   { type: string, example: Sharma }
        jobTitle:   { type: string, example: Senior Business Analyst }
        department: { type: string, example: Product Engineering }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:    { type: string, format: email }
        password: { type: string }
    AuthUser:
      type: object
      properties:
        id:              { type: string, format: uuid }
        email:           { type: string, format: email }
        firstName:       { type: string }
        lastName:        { type: string }
        role:            { $ref: '#/components/schemas/UserRole' }
        isEmailVerified: { type: boolean }
        avatarUrl:       { type: string, nullable: true }
        jobTitle:        { type: string, nullable: true }
        department:      { type: string, nullable: true }
        createdAt:       { type: string, format: date-time }
    LoginResponse:
      type: object
      properties:
        success:     { type: boolean, example: true }
        data:
          type: object
          properties:
            accessToken: { type: string, description: JWT Bearer token (15 min) }
            tokenType:   { type: string, example: Bearer }
            expiresIn:   { type: integer, example: 900, description: Seconds }
            user:        { $ref: '#/components/schemas/AuthUser' }
    RefreshResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            accessToken: { type: string }
            tokenType:   { type: string, example: Bearer }
            expiresIn:   { type: integer, example: 900 }

    # ── Requirement Schemas ───────────────────────────────────────────────────
...
    HistoryItem:
      type: object
      properties:
        analysisId:       { type: string, format: uuid }
        requirementId:    { type: string, format: uuid }
        requirementTitle: { type: string }
        projectId:        { type: string, format: uuid }
        projectName:      { type: string }
        status:           { $ref: '#/components/schemas/AnalysisStatus' }
        aiProvider:       { $ref: '#/components/schemas/AIProvider' }
        aiModel:          { type: string }
        complexityLevel:  { $ref: '#/components/schemas/ComplexityLevel' }
        riskCount:        { type: integer, nullable: true }
        storyCount:       { type: integer, nullable: true }
        tokensTotal:      { type: integer, nullable: true }
        costUsd:          { type: number, nullable: true }
        triggeredBy:      { type: string, format: uuid }
        completedAt:      { type: string, format: date-time, nullable: true }
    HistoryListResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: array
          items: { $ref: '#/components/schemas/HistoryItem' }
        meta: { $ref: '#/components/schemas/PaginationMeta' }

    # ── Export Schema ─────────────────────────────────────────────────────────
...
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        "201":
          description: Account created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/AuthUser' }
        "409": { $ref: '#/components/responses/Conflict' }
        "422": { $ref: '#/components/responses/ValidationError' }
  /auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate and receive JWT access token
      description: |
        Returns an access token (15 min) in the response body and sets a
        `refreshToken` httpOnly cookie (7 days) for token rotation.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
            example:
              email: ananya@company.com
              password: "MyP@ssw0rd!"
      responses:
        "200":
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
                example: refreshToken=<uuid>; HttpOnly; Secure; SameSite=Strict; Path=/api/v1/auth/refresh
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "422": { $ref: '#/components/responses/ValidationError' }
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Rotate refresh token and issue new access token
      description: Reads the `refreshToken` httpOnly cookie. Rotates it on every use.
      responses:
        "200":
          description: New access token issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RefreshResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Invalidate refresh token and clear session cookie
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Logged out successfully
        "401": { $ref: '#/components/responses/Unauthorized' }
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get current authenticated user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/AuthUser' }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Request a password reset email
      description: Always returns 200 to prevent email enumeration attacks.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "200":
          description: If email exists, a reset link has been sent
  /auth/reset-password:
    post:
      tags: [Authentication]
      summary: Reset password using a valid reset token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token:    { type: string }
                password: { type: string, minLength: 8 }
      responses:
        "200":
          description: Password reset successfully
        "400":
          description: Invalid or expired reset token
  /auth/change-password:
    post:
      tags: [Authentication]
      summary: Change password for the authenticated user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string }
                newPassword:     { type: string, minLength: 8 }
      responses:
        "204":
          description: Password changed successfully
        "401": { $ref: '#/components/responses/Unauthorized' }

  # ── Requirements ─────────────────────────────────────────────────────────
...
  /history:
    get:
      tags: [History]
      summary: Paginated analysis history across all user's projects
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: projectId
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/AnalysisStatus' }
        - name: provider
          in: query
          schema: { $ref: '#/components/schemas/AIProvider' }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [completedAt, createdAt, tokensTotal, costUsd]
            default: completedAt
        - name: sortDir
          in: query
          schema: { type: string, enum: [asc, desc], default: desc }
      responses:
        "200":
          description: Paginated history list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HistoryListResponse' }
  /history/saved:
    get:
      tags: [History]
      summary: Get user's bookmarked analyses
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        "200":
          description: Saved analysis list
  /history/saved/{analysisId}:
    parameters:
      - $ref: '#/components/parameters/AnalysisId'
    delete:
      tags: [History]
      summary: Remove an analysis from saved bookmarks
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Bookmark removed

  # ── Export ────────────────────────────────────────────────────────────────
...